Removed long time ago obsolete quoted-printable decoding code from singlefile() which was used for daily build log files named 'inbox*log'.
At least since 2007, received autobuild files are first processed by the inpipe
script, and once decoded if needed, stored in an individual build log file named
'build*log' and additionally in a daily build log file named 'inbox*log'.

In this way, if decoding is required for a received autobuild file it is only
done once to avoid loading the server needlessly.