Make unresponsive test server get counted as a warning.
In a perfect world a torture test that left the server unresponsive should
be counted as a failed torture test given that at some point is has been
following unexpected code paths; no matter if it is a server problem or a
test tool problem.

This avoids modifying runtests.pl now and still notice these failures.