summarize.pl: handle non-autoconf builds in the buildcode
Such builds (including Netware and some Windows) are configured
externally using make arguments or other means that are not visible
in the test log header. To make those build codes unique, include
the contents of curlbuild.h and curl_config.h (or config-win32.h)
in the hash. This is undesirable for autoconf builds as it makes
the code sensitive to uninteresting internal configure details.