vulntable: generate and link to specific-version pages
The main table now uses rowspan for the height of each vulnerability,
which shrunk the total size of the HTML used for the table to a mere 20%
of the original size.

The version numbers in the table now links to single version
vulnerability summaries.

Each specific-version page details vulnerabilities for that specific
version - all generated by the vuln.pm data and the new
_singlevuln.templ template. Also, now link to the advisory page for
details on each problem instead of the security.html page and an anchor.