adv_20170419: TLS session resumption client cert bypass (again)
libcurl would attempt to resume a TLS session even if the client
certificate had changed. That is unacceptable since a server by
specification is allowed to skip the client certificate check on resume,
and may instead use the old identity which was established by the
previous certificate (or no certificate).

CVE-2017-7468

Bug: https://curl.haxx.se/docs/adv_20170419.html