libcurl/c/Makefile: fix option man page generation
The build process has an intermediate step where it creates an include
file that is then included by a template file. Anything that then looks
like a (C) comment will be handled as a comment in the second step and
thus we now replace '*' with its HTML entity version so that it will
look right but won't be treated as a comment.

Bug: https://github.com/bagder/roffit/issues/23
Reported-by: Sai Ram Kunala