vuln: add area and "C issue" to each vulnerability
To help facilitate discussion and analyses on the existing set of curl
vulnerabilities (within and outside of the project), this change adds a
single-word area for each issue as well as a "C mistake" for those
problems that were due to a C error.

This data *could* be maintained separately but I think by adding these
to the main vuln table we can reach a single view on how to "categorize"
each flaw and we can make sure that we do the same work on future
vulnerabilities as well.