dl/remcheck.pl: Add --ssl for checking ftp sites
This option is ignored on http sites and succeeds on all the ftp sites
checked as of this moment. Not all of them support encryption, but it
gets silently ignored on those and then we're not any worse off than
today. We can't add --ssl-reqd because of those servers, so this ends up
being more a privacy enhancement than a security one. Also, rename an
(unused) variable to reflect a time span change.