curlopt2href: linkify "bare" libcurl symbols in man pages
Extends the MAN2HTML macro by also checking for curl symbols in the text
that appears to not be linked already and links them to their
corresponding web page.

This primarily adds links to the example section and synopsis sections
of the libcurl man pages as they typically cannot highlight the
identifiers using nroff.

Closes #263