docs/optionswhen.html: add
A new page listing all curl versions in a reverse chronological order
and the command line options added in that version. With links to the
changelog on the version number and to the relevant man page section on
the command line option name.

Bonus: the page formerly sometimes called "manual" is now called
tutorial everywhere, but the URL remains called "manual.html" at least
for now.

Closes #303