docs/security: show the severity filter status
- Add a sentence above the vulnerability table that lets the user know
  whether or not the table has a severity filter applied.

Example: (The table below has been filtered to show High+ severity)

This is a follow-up to 90850a74 which added generated security pages
for different severity levels. security-m.html (Medium),
security-h.html (High) and security-c.html (Critical).

When I google search for `curl cves` the second result is a link
to security-h.html. It may not be immediately obvious that the table is
filtered and not showing all vulnerabilities.

Closes #305