CVE-2023-38545.md: add additional info
RyotaK found a way to exploit the bug by crafting a hostname so that the
SOCKS connection completes successfully but to a different hostname and
port, and possibly arbitrary data is sent to that different host before
the "expected" request is sent on that same connection.

The impact is limited because control characters and null are not
allowed in the hostname.

Ref: https://hackerone.com/ryotak?type=user

Closes https://github.com/curl/curl-www/pull/308