GHA: use `pyspelling` directly, setup yearly Dependabot updates (#486)
To avoid these extra layers when using `pyspelling`:
- a GitHub Action
- a docker container image
- Docker Hub

The cost is a slower install step. Benefits are 3x faster, parallel
spellcheck, being in sync with the same logic in curl/curl, and direct
pin and bump for the pip package. Mostly the latters, because this is
a rarely run CI job, and speed isn't critical.

pyspelling is configured to use aspell and aspell-en. These packages
are not seeing many updates anymore, and after this patch aspell is
bumped to its latest minor revision 0.60.8.1 (from 0.60.8):
https://sources.debian.org/src/aspell/
https://sources.debian.org/src/aspell-en/